requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==4.2.2
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
attrs==24.2.0
azure-core==1.32.0
azure-identity==1.19.0
azure-storage-blob==12.24.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
captcha==0.6.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
coverage==7.6.4
cryptography==44.0.0
debugpy==1.8.7
decorator==5.1.1
distro==1.9.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==2.1.0
extra-streamlit-components==0.1.71
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.28.0
isodate==0.7.2
jedi==0.19.1
jellyfish==1.1.0
Jinja2==3.1.4
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.4
langchain-core==0.3.12
langchain-text-splitters==0.3.0
langsmith==0.1.137
Levenshtein==0.26.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
msal==1.31.1
msal-extensions==1.2.0
multidict==6.1.0
narwhals==1.9.4
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.52.0
openpyxl==3.1.5
orjson==3.10.10
packaging==24.1
pandas==2.2.3
parso==0.8.4
pefile==2023.2.7
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
portalocker==2.10.1
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.2
psutil==6.1.0
pure_eval==0.2.3
pyaml==24.9.0
pyarrow==17.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
PyJWT==2.9.0
pytest==8.3.3
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.10.0
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.2
rpds-py==0.20.0
setuptools==75.3.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
streamlit==1.39.0
streamlit-aggrid==1.0.5
streamlit-authenticator==0.4.1
tenacity==9.0.0
toml==0.10.2
toolz==1.0.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==5.0.3
wcwidth==0.2.13
XlsxWriter==3.2.0
yarl==1.16.0